attempting to simplify modal
